chore(Root): adapt tests
