+++ .github/workflows/ci.yml
@@ -1,4 +1,4 @@
-name: CI | Tests & release
+name: CI
 
 on:
   push:
@@ -8,9 +8,10 @@ on:
 
 jobs:
   check-skip:
+    name: Skip redundant builds
     runs-on: ubuntu-latest
     steps:
-      - name: Cancel Previous Redundant Builds
+      - name: Cancel previous redundant builds
         uses: styfle/cancel-workflow-action@0.6.0
         with:
           access_token: ${{ github.token }}
