+++ action.yml
@@ -29,7 +29,7 @@ runs:
         urls: ${{inputs.urls}}
     - name: Format Lighthouse score
       id: format_lhci_score
-      run: node ${{ github.action_path }}/scripts/generateCommentFromResults.js ${{ steps.lhci.outputs.manifest }}
+      run: node ${{ github.action_path }}/scripts/formatScore.js ${{ steps.lhci.outputs.manifest }}
       shell: bash
     - name: Add comment to PR
       id: comment_to_pr
